refactor: TiemtableView가 Columns, AllDay로 분리되기 위해 TimetableView -> TimetableColumns로 변경 (#39)
